Initial commit of a Picard wrapper.
Copied code from the galaxy library:

https://bitbucket.org/galaxy/galaxy-dist/src/26920e20157f/tools/picard/picard_wrapper.py

Began an extensive overhaul, starting with using argparse and
subparsers.

The new structure will use the subparsers to run the subcommands, so a
function for each subcommand will be implemented. A setup() and
cleanup() method will wrap around each sub-function (this may not be
necessary if/when I determine what all the current setup code is
doing; just copied things from the top of the main function to make
sure it's working). Currently, I'm using a decorator for this.

I haven't been able to test, since I currently don't have any SAM/BAM
files!